feat: CXSPA-3859 move cpq converters and model to common folder as they will be used for both rest and occ services
