+++ feature-libs/product-configurator/rulebased/cpq/common/converters/cpq-configurator-normalizer-utils.service.spec.ts
