+++ feature-libs/product-configurator/rulebased/cpq/common/converters/cpq-configurator-overview-normalizer.ts
