+++ feature-libs/product-configurator/rulebased/cpq/common/converters/cpq-configurator-serializer.spec.ts
