+++ feature-libs/product-configurator/rulebased/cpq/common/converters/cpq-configurator-serializer.ts
@@ -8,7 +8,7 @@ import { Injectable } from '@angular/core';
 import { Converter } from '@spartacus/core';
 import { Configurator } from '@spartacus/product-configurator/rulebased';
 import { Cpq } from '../cpq.models';
-import { CpqConfiguratorUtils } from './../cpq-configurator-utils';
+import { CpqConfiguratorUtils } from '../../rest/cpq-configurator-utils';
 
 const VALUE_SEPARATOR = ',';
