+++ feature-libs/product-configurator/rulebased/cpq/common/converters/cpq-configurator-value-serializer.spec.ts
