+++ feature-libs/product-configurator/rulebased/cpq/public_api.ts
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+export * from './common/index';
 export * from './occ/index';
 export * from './rest/index';
 export * from './rulebased-cpq-configurator.module';
