+++ feature-libs/product-configurator/rulebased/cpq/rest/index.ts
@@ -4,11 +4,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './converters/index';
 export * from './cpq-configurator-endpoint.config';
 export * from './cpq-configurator-endpoint.service';
 export * from './cpq-configurator-rest.adapter';
 export * from './cpq-configurator-rest.module';
 export * from './cpq-configurator-rest.service';
 export * from './cpq-configurator-utils';
-export * from './cpq.models';
